+++ AutomatedLab.Recipe/Public/Save-LabRecipe.ps1
@@ -1,4 +1,4 @@
-function Save-LabRecipe
+﻿function Save-LabRecipe
 {
     [CmdletBinding(SupportsShouldProcess)]
     param